requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.12.0
